+++ DCM Framework/DCMPixelDataAttribute.m
@@ -78,7 +78,7 @@ void SwapLongs( register vector unsigned int *unaligned_input, register long siz
 {
 	 register long i = size / 4;
 	 register vector unsigned char identity = vec_lvsl(0, (int*) NULL );
-	 register vector unsigned char byteSwapLongs = vec_xor( identity, vec_splat_u8(sizeof( long )- 1 ) );
+	 register vector unsigned char byteSwapLongs = vec_xor( identity, vec_splat_u8(sizeof( int )- 1 ) );
 	 while(i-- > 0)
 	 {
 	 *unaligned_input++ = vec_perm( *unaligned_input, *unaligned_input, byteSwapLongs);
